+++ src/KubernetesWrapper/src/Controllers/OperationalLogsController.cs
@@ -25,7 +25,7 @@ namespace KubernetesWrapper.Controllers
         /// <returns>The list of operational logs</returns>
         [HttpGet]
         [EnableCors]
-        public async Task<ActionResult<IEnumerable<Log>>> GetLogs(string app = null, double take = 50, double time = 24)
+        public async Task<ActionResult<IEnumerable<Log>>> GetLogs(string app = null, int take = 50, double time = 24)
         {
             var logs = await operationalLogsService.GetLogs(app, take, time);
             return Ok(logs);
